requirements.txt
aiofiles==23.2.1
aiosmtplib==2.0.2
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
configobj==5.0.8
configparser==6.0.1
dnspython==2.6.1
email-validator==2.1.1
etelemetry==0.3.1
exceptiongroup==1.2.0
fastapi==0.110.0
fastapi-jwt-auth==0.5.0
fastapi-mail==1.4.1
filelock==3.13.1
fitz==0.0.1.dev2
frontend==0.0.3
h11==0.14.0
httplib2==0.22.0
idna==3.6
importlib-resources==6.1.3
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
looseversion==1.3.0
lxml==5.1.0
MarkupSafe==2.1.5
networkx==3.1
nibabel==5.2.1
nipype==1.8.6
numpy==1.24.4
packaging==24.0
pandas==2.0.3
pathlib==1.0.1
pillow==10.2.0
pip==24.0
prov==2.0.0
pydantic==2.6.3
pydantic-core==2.16.3
pydantic-settings==2.2.1
pydot==2.0.0
PyJWT==1.7.1
pymongo==4.6.2
PyMuPDF==1.23.26
PyMuPDFb==1.23.22
pyparsing==3.1.2
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytils==0.4.1
pytz==2024.1
pyxnat==1.6.2
rdflib==7.0.0
requests==2.31.0
scipy==1.10.1
setuptools==41.2.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
tools==0.1.9
traits==6.3.2
typing-extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.28.0
zipp==3.17.0
